requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
pyupbit
pandas
websockets
